fix kafka template for message
