+++ src/main/java/leyans/RidersHub/Config/KafkaConfig/KafkaREST/Consumer/RiderConsumer.java
@@ -10,8 +10,8 @@ public class RiderConsumer {
 
         @KafkaListener(topics = {"car", "motor", "bike"}, groupId = "rider-group")
         public void consume(ConsumerRecord<String, String> record) {
-            String riderType = record.topic();  // The topic is now the rider type
-            String message = record.value();  // Get the actual message
+            String riderType = record.topic();
+            String message = record.value();
 
             System.out.println("Received message: " + message + " for RiderType: " + riderType);
         }
