+++ src/main/java/leyans/RidersHub/Service/RiderService.java
@@ -6,6 +6,7 @@ import leyans.RidersHub.Repository.RiderTypeRepository;
 import leyans.RidersHub.model.Rider;
 import leyans.RidersHub.model.RiderType;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 import leyans.RidersHub.Repository.RiderRepository;
 
@@ -20,6 +21,9 @@ public class RiderService {
 
     @Autowired
     private final RiderTypeRepository riderTypeRepository;
+
+    @Autowired
+    KafkaTemplate<String, String> kafkaTemplate;
 
     public RiderService(RiderRepository riderRepository, RiderTypeRepository riderTypeRepository) {
         this.riderRepository = riderRepository;
@@ -43,6 +47,15 @@ public class RiderService {
         newRider.setEnabled(enabled);
         newRider.setRiderType(riderTypeName);
         return riderRepository.save(newRider);
+
+    }
+
+    public void sendMessage(String riderType, String message) {
+
+
+        System.out.println("Sending message to topic: " + riderType);
+        kafkaTemplate.send(riderType, message);
+        System.out.println("Sent message: " + message + " to RiderType: " + riderType);
 
     }
