+++ src/main/java/leyans/RidersHub/Service/RidesService.java
@@ -2,6 +2,7 @@ package leyans.RidersHub.Service;
 
 
 import jakarta.transaction.Transactional;
+import leyans.RidersHub.DTO.RideResponseDTO;
 import leyans.RidersHub.DTO.RidesDTO;
 import leyans.RidersHub.Repository.RiderRepository;
 import leyans.RidersHub.Repository.RiderTypeRepository;
@@ -12,6 +13,7 @@ import leyans.RidersHub.model.Rides;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Point;
+import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -22,66 +24,21 @@ public class RidesService {
     private final RidesRepository ridesRepository;
     private final RiderRepository riderRepository;
     private final RiderTypeRepository riderTypeRepository;
-    private final LocationService   locationService;
+    private final KafkaTemplate<Object, RidesDTO> kafkaTemplate;
 
 
     private final GeometryFactory geometryFactory = new GeometryFactory();
 
 
-    public RidesService(RiderRepository riderRepository, RiderTypeRepository riderTypeRepository, LocationService locationService, RidesRepository ridesRepository) {
+    public RidesService(RiderRepository riderRepository, RiderTypeRepository riderTypeRepository, RidesRepository ridesRepository, KafkaTemplate<Object, RidesDTO> kafkaTemplate) {
         this.riderRepository = riderRepository;
         this.riderTypeRepository = riderTypeRepository;
-        this.locationService = locationService;
         this.ridesRepository = ridesRepository;
+        this.kafkaTemplate = kafkaTemplate;
     }
 
-//    @Transactional
-//    public Rides createRides(String username, String ridesName, String
-//            locationName, List<String> riderTypeNames, Integer distance, String startingPoint, Date date, double latitude, double longitude) {
-//
-//
-//
-//        Rider rider = riderRepository.findByUsername(username);
-//
-//
-//        Point coordinates = geometryFactory.createPoint(new Coordinate(longitude, latitude));
-//        Locations location = new Locations(rider, locationName, coordinates);
-//        location = locationRepository.save(location);
-//
-//        String pointStr = coordinates.getX() + "," + coordinates.getY();
-//        LocationDTO locationDTO = new LocationDTO(username, locationName, pointStr);
-//
-//        kafkaProducer.sendLocationUpdate( locationDTO);
-//
-//        // Fetch RiderType objects from the database
-//        List<RiderType> riderTypes = riderTypeNames.stream()
-//                .map(name -> riderTypeRepository.findByRiderType(name))
-//                .filter(Objects::nonNull) // Avoid null values if no match is found
-//                .collect(Collectors.toList());
-//
-//        if (riderTypes.isEmpty()) {
-//            throw new IllegalArgumentException("No valid RiderType found for given names.");
-//        }
-//
-//
-//
-//        Rides newRides = new Rides();
-//        newRides.setCoordinates(coordinates);
-//        newRides.setLocationName(locationName);
-//        newRides.setRidesName(ridesName);
-//        newRides.setUsername(rider);
-//        newRides.setDistance(distance);
-//        newRides.setStartingPoint(startingPoint);
-//        newRides.setDate(date);
-//        newRides.setRiderTypes(riderTypes); // Set fetched rider types
-//
-//
-//        return ridesRepository.save(newRides);
-//    }
-
-
     @Transactional
-    public Rides createRide(String username, String ridesName, String
+    public RideResponseDTO createRide(String username, String ridesName, String
             locationName, String riderType, Integer distance, String startingPoint, Date date, double latitude, double longitude) {
 
 
@@ -91,9 +48,11 @@ public class RidesService {
         Point coordinates = geometryFactory.createPoint(new Coordinate(longitude, latitude));
 
         String pointStr = coordinates.getX() + "," + coordinates.getY();
-        RidesDTO ridesDTO = new RidesDTO(username, locationName, pointStr, ridesName,
+
+        RidesDTO ridesDTO = new RidesDTO(username, locationName, ridesName,
                 riderType, distance, startingPoint, date
         );
+
 
         Rides newRides = new Rides();
         newRides.setCoordinates(coordinates);
@@ -104,9 +63,11 @@ public class RidesService {
         newRides.setStartingPoint(startingPoint);
         newRides.setDate(date);
         newRides.setRiderType(newRiderType);
-        return ridesRepository.save(newRides);
+        newRides = ridesRepository.save(newRides);
 
+        kafkaTemplate.send("rides-group", );
 
+        return new RideResponseDTO(ridesDTO, pointStr);
     }
 }
